Fix calcul palettes (entier) + Entrepôt 3D fonctionnel à grande capacité
- calculer_occupation : palettes = (int) ceil(Disponible / Qté par palette).
  Résultat toujours entier (ex. 2 / 45 => 1, jamais 0,044).
- Entrepôt 3D réécrit en rendu instancié (InstancedMesh) : supporte des
  capacités de plusieurs milliers d'emplacements (11 600) sans figer la page.
  Palettes réparties sur toute l'emprise, bascule couleur catégorie/alerte,
  filtres et survol conservés ; boîtes libres seulement si peu nombreuses.

+++ rapport-stock/includes/report.php
@@ -460,9 +460,11 @@ function calculer_occupation(array $lignes, float $capacite): array
         $sansPalette = ($parPal <= 0);
         if ($sansPalette) {
             $nbSansPalette++;
-            $palettes = 0.0;
+            $palettes = 0;
         } else {
-            $palettes = $dispo > 0 ? (float) ceil($dispo / $parPal) : 0.0;
+            // Une palette entamée occupe un emplacement complet : arrondi au supérieur.
+            // Résultat TOUJOURS entier (ex. Disponible 2 / 45 par palette => 1, jamais 0,044).
+            $palettes = $dispo > 0 ? (int) ceil($dispo / $parPal) : 0;
         }
         $occupees += $palettes;
